refactor: put schema structs in one file
