+++ src/postgres/mod.rs
@@ -1,11 +1,9 @@
-mod coordinator;
-mod data;
 mod rows;
+mod schemas;
 
 // Re-export types from submodules.
-pub use coordinator::*;
-pub use data::*;
 pub use rows::*;
+pub use schemas::*;
 
 // Import common dependencies below.
 use crate::types::Metric;
@@ -14,50 +12,6 @@ use sqlx::PgConnection;
 
 type SchemaName = Box<str>;
 type TableName = Box<str>;
-
-/// Trait of a node schema in Postgres database.
-///
-/// The schema of a coordinator node and a data node are mostly different. This
-/// trait defines the common methods for both types of nodes.
-#[async_trait]
-pub trait NodeSchema {
-    /// Return the schema name of the node.
-    fn schema(&self) -> SchemaName;
-
-    /// Return the table name storing cluster data.
-    fn cluster_table(&self) -> TableName {
-        format!("{}.clusters", self.schema()).into_boxed_str()
-    }
-
-    /// Create a new schema belonging to a node in the database.
-    async fn create_schema(&self, connection: &mut PgConnection) {
-        sqlx::query(&format!("CREATE SCHEMA IF NOT EXISTS {}", self.schema()))
-            .execute(connection)
-            .await
-            .expect("Failed to create the schema");
-    }
-
-    /// Create all tables required by the node.
-    async fn create_all_tables(&self, connection: &mut PgConnection);
-
-    /// Create a table to store cluster data.
-    ///
-    /// Columns:
-    /// - id: Cluster ID.
-    /// - centroid: Centroid vector of the cluster.
-    async fn create_cluster_table(&self, connection: &mut PgConnection) {
-        let table = self.cluster_table();
-        sqlx::query(&format!(
-            "CREATE TABLE IF NOT EXISTS {table} (
-                id UUID PRIMARY KEY DEFAULT gen_random_uuid(),
-                centroid BYTEA NOT NULL
-            )"
-        ))
-        .execute(connection)
-        .await
-        .expect("Failed to create cluster table");
-    }
-}
 
 #[cfg(test)]
 pub mod test_utils {
